+++ cbl-mariner/oval/testdata/golden/cbl-mariner/oval/1.0/definitions/2008/CVE-2008-3914.json
