+++ cbl-mariner/oval/testdata/golden/cbl-mariner/oval/1.0/definitions/2018/CVE-2018-25012.json
