+++ cbl-mariner/oval/testdata/golden/cbl-mariner/oval/1.0/objects/objects.json
