+++ cbl-mariner/oval/testdata/golden/cbl-mariner/oval/2.0/definitions/2014/CVE-2014-8139.json
