+++ cbl-mariner/oval/testdata/golden/cbl-mariner/oval/2.0/definitions/2021/CVE-2021-39924.json
