+++ cbl-mariner/oval/testdata/golden/cbl-mariner/oval/2.0/definitions/2022/CVE-2022-21309.json
