+++ cbl-mariner/oval/testdata/golden/cbl-mariner/oval/2.0/objects/objects.json
