+++ cbl-mariner/oval/testdata/happy/cbl-mariner-1.0-oval.xml
