+++ cbl-mariner/oval/testdata/happy/cbl-mariner-2.0-preview-oval.xml
